+++ __tests__/index.js
@@ -4,15 +4,6 @@ import { mount, shallow } from 'enzyme';
 import Modal from '../src';
 
 describe('Modal', () => {
-  it('Should have sane default props', () => {
-    const wrapper = mount(<Modal />);
-
-    expect(wrapper.props().closeOnEsc).toEqual(false);
-    expect(wrapper.props().closeOnOutsideClick).toEqual(false);
-    expect(wrapper.props().open).toEqual(true);
-    expect(wrapper.state('open')).toEqual(true);
-  });
-
   it('Should obey open=false prop', () => {
     const wrapper = mount(<Modal open={false} />);