+++ config/jest/setup.js
@@ -1,3 +1,5 @@
+import '@babel/register';
+
 import Enzyme, { shallow, render, mount } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';