+++ src/components/Backdrop.ts
@@ -1,0 +1,7 @@
+import styled, { StyledComponentClass } from 'styled-components';
+
+const Backdrop: StyledComponentClass<any, any> = styled.div`
+  background: rgba(0, 0, 0, 0.32);
+`;
+
+export default Backdrop;