+++ src/components/Overscroll.ts
@@ -1,0 +1,12 @@
+import styled, { StyledComponentClass } from 'styled-components';
+
+const Overscroll: StyledComponentClass<any, any> = styled.div`
+  align-items: center;
+  display: flex;
+  justify-content: center;
+  min-height: 100%;
+  position: relative;
+  width: 100%;
+`;
+
+export default Overscroll;