+++ src/index.js
@@ -1,5 +1,0 @@
-import Modal from './components/Modal';
-
-export { Portal, collectPortals, flushPortals } from './components/Portal';
-
-export default Modal;