+++ src/index.ts
@@ -1,0 +1,9 @@
+import Modal from './components/Modal';
+
+export {
+  default as Portal,
+  collectPortals,
+  flushPortals
+} from './components/Portal';
+
+export default Modal;