+++ src/styles/index.js
@@ -1,2 +1,0 @@
-export { default as container } from './container';
-export { default as backdrop } from './backdrop';