+++ src/utils/has-dom.ts
@@ -1,0 +1,1 @@
+export default () => window && window.document && window.document.createElement;