+++ stories/03-styled.tsx
@@ -1,20 +1,20 @@
 import React, { Fragment } from 'react';
 import { storiesOf } from '@storybook/react';
-import styled from 'styled-components';
+import styled, { StyledComponentClass } from 'styled-components';
 
-import Modal from '../src';
+import Modal from 'styled-modal';
 
-const Overlay = styled.div`
+const Overlay: StyledComponentClass<any, any> = styled.div`
   background-color: rgba(242, 242, 242, 0.96);
-  padding: 32px 16px;
 `;
 
-const Container = styled.article`
+const Container: StyledComponentClass<any, any> = styled.article`
   background-color: white;
   border-radius: 8px;
   box-shadow: 0 4px 64px rgba(0, 0, 0, 0.08), 0 1px 4px rgba(0, 0, 0, 0.08);
   max-width: 480px;
-  padding: 16px;
+  min-height: 120vh;
+  margin: 16px;
   position: relative;
   text-align: center;
   width: 100%;