+++ stories/05-portal.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+
+import { Portal } from 'styled-modal';
+
+storiesOf('portal-modal', module).add('Portal with custom target', () => {
+  const target = document.getElementById('error-display');
+
+  return (
+    <React.Fragment>
+      <Portal target={target}>
+        <p>This text is portaled to Storybook's default #error-display</p>
+      </Portal>
+    </React.Fragment>
+  );
+});