+++ typings/dom-focus-lock.d.ts
@@ -1,0 +1,8 @@
+declare module 'dom-focus-lock/umd' {
+  export function on(element?: HTMLElement) {
+    return;
+  }
+  export function off(element?: HTMLElement) {
+    return;
+  }
+}