+++ typings/no-scroll.d.ts
@@ -1,0 +1,8 @@
+declare module 'no-scroll' {
+  export function on() {
+    return;
+  }
+  export function off() {
+    return;
+  }
+}