+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2018-03-30
+### Changed
+- Rewrite in Babel 7 / Typescript
+- Update for React 16.3.0
+- `<Portal />` can be render to whatever dom element by supplying it in the `target` prop
+
 ## [1.1.0] - 2018-03-10
 ### Added
 - Pass all other props supplied to `<Modal />` to the container component. This allow you to set `aria-labelledby` and `aria-describedby`, for example